fix: mobile formula display
